+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/MoshiChatApi.kt
@@ -36,7 +36,6 @@ import io.getstream.chat.android.client.api2.mapping.toDomain
 import io.getstream.chat.android.client.api2.mapping.toDto
 import io.getstream.chat.android.client.api2.model.dto.ChatEventDto
 import io.getstream.chat.android.client.api2.model.dto.DeviceDto
-import io.getstream.chat.android.client.api2.model.dto.DownstreamChannelUserRead
 import io.getstream.chat.android.client.api2.model.dto.DownstreamMemberDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamMessageDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamReactionDto
@@ -81,6 +80,7 @@ import io.getstream.chat.android.client.api2.model.response.TranslateMessageRequ
 import io.getstream.chat.android.client.api2.model.response.VideoCallTokenResponse
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.extensions.enrichWithCid
+import io.getstream.chat.android.client.extensions.syncUnreadCountWithReads
 import io.getstream.chat.android.client.helpers.CallPostponeHelper
 import io.getstream.chat.android.client.parser.toMap
 import io.getstream.chat.android.client.scope.UserScope
@@ -712,15 +712,14 @@ constructor(
         return response.channel.toDomain().let { channel ->
             channel.copy(
                 watcherCount = response.watcher_count,
-                read = response.read.map(DownstreamChannelUserRead::toDomain),
+                read = response.read.map { it.toDomain(channel.lastMessageAt ?: it.last_read) },
                 members = response.members.map(DownstreamMemberDto::toDomain),
                 membership = response.membership?.toDomain(),
                 messages = response.messages.map { it.toDomain().enrichWithCid(channel.cid) },
                 watchers = response.watchers.map(DownstreamUserDto::toDomain),
                 hidden = response.hidden,
                 hiddenMessagesBefore = response.hide_messages_before,
-                unreadCount = response.read.firstOrNull { it.user.id == userId }?.unread_messages,
-            )
+            ).syncUnreadCountWithReads()
         }
     }