+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/channels/list/ChannelItem.kt
@@ -45,6 +45,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import io.getstream.chat.android.client.extensions.currentUserUnreadCount
 import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.previewdata.PreviewChannelData
 import io.getstream.chat.android.compose.previewdata.PreviewUserData
@@ -249,9 +250,9 @@ internal fun RowScope.DefaultChannelItemTrailingContent(
                 .align(Alignment.Bottom),
             horizontalAlignment = Alignment.End,
         ) {
-            val unreadCount = channel.unreadCount
+            val unreadCount = channel.currentUserUnreadCount
 
-            if (unreadCount != null && unreadCount > 0) {
+            if (unreadCount > 0) {
                 UnreadCountIndicator(
                     modifier = Modifier.padding(bottom = 4.dp),
                     unreadCount = unreadCount,