+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/channels/list/adapter/internal/ChannelListItemDiffCallback.kt
@@ -17,6 +17,7 @@
 package io.getstream.chat.android.ui.feature.channels.list.adapter.internal
 
 import androidx.recyclerview.widget.DiffUtil
+import io.getstream.chat.android.client.extensions.currentUserUnreadCount
 import io.getstream.chat.android.client.extensions.getMembersExcludingCurrent
 import io.getstream.chat.android.ui.common.extensions.internal.cast
 import io.getstream.chat.android.ui.common.extensions.internal.safeCast
@@ -67,7 +68,7 @@ internal object ChannelListItemDiffCallback : DiffUtil.ItemCallback<ChannelListI
             usersChanged = usersChanged,
             readStateChanged = channel.read != other.channel.read,
             lastMessageChanged = channel.getLastMessage() != other.channel.getLastMessage(),
-            unreadCountChanged = channel.unreadCount != other.channel.unreadCount && other.channel.unreadCount != null,
+            unreadCountChanged = channel.currentUserUnreadCount != other.channel.currentUserUnreadCount,
             extraDataChanged = channel.extraData != other.channel.extraData,
             typingUsersChanged = typingUsers != other.typingUsers,
         )